+++ CHANGELOG.md
@@ -7,12 +7,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Added
 - Added Trustee field to qrz command for club callsigns.
+- Added alias for `ae7q call` command (`ae7q c`).
+- Added ae7q lookup by FRN and Licensee ID, and for trustee records (`ae7q frn, licensee, trustee`).
 ### Changed
 - Changelog command to accept a version as argument.
 - The qrz command can now link to a QRZ page instead of embedding the data with the `--link` flag.
+- All currently-available pools can now be accessed by the `hamstudy` command.
+- The `hamstudy` command now uses the syntax `?hamstudy <country> <pool>`.
+- Replaced `hamstudyanswer` command with answering by reaction.
 ### Fixed
 - Fixed ditto marks (") appearing in the ae7q call command.
 - Fixed issue where incorrect table was parsed in ae7q call command.
+- Fixed warning emoji reaction on messages starting with "??".
 
 
 ## [2.1.0] - 2020-01-04